Переписан первый спайдер
